1.0.1 commit
Reworked a lot of the back end to function with/without the database
(the user never needs to log in, and use the database features, but
they're there).

All of the old XML file interaction stuff should make more sense yet,
but there's still things that I need to fix in order for them to be less
messy (ex: the AddMovie function was basically useless, and was causing
a lot of slowdown on large trasactions 300kb took about 3 minutes to
download).